Hand fix and warning check (#1302)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Remove tavern and guest key from hand so their keyring has 10 keys and
actually gets them all.
Add a stack trace for keyrings with over 10 keys. (If there is a storage
specific way of checking ill change it to that)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Hand doesn't get the manor key because he has too many keys.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

+++ code/game/objects/items/keyrings.dm
@@ -19,10 +19,15 @@
 
 /obj/item/storage/keyring/Initialize()
 	. = ..()
-	for(var/X in keys)
+	if(!length(keys))
+		return
+	if(length(keys) > 10)
+		stack_trace("Keyring [src] has too many keys and the list will get cut short!")
+	for(var/X as anything in keys)
 		var/obj/item/key/new_key = new X(loc)
 		if(!SEND_SIGNAL(src, COMSIG_TRY_STORAGE_INSERT, new_key, null, TRUE, FALSE))
 			qdel(new_key)
+		LAZYREMOVE(keys, X)
 
 	update_icon()
 	update_desc()
@@ -50,12 +55,12 @@
 			icon_state = "keyring5"
 
 /obj/item/storage/keyring/proc/update_desc()
-	if(contents.len)
-		desc = span_info("Holds \Roman[contents.len] key\s, including:")
-		for(var/obj/item/key/KE in contents)
-			desc += span_info("\n- [KE.name ? "\A [KE.name]." : "An unknown key."]")
-	else
-		desc = ""
+	if(!length(contents))
+		desc = initial(desc)
+		return
+	desc = span_info("Holds \Roman[length(contents)] key\s, including:")
+	for(var/obj/item/key/KE in contents)
+		desc += span_info("\n- [KE.name ? "\A [KE.name]." : "An unknown key."]")
 
 /obj/item/storage/keyring/Entered(atom/movable/arrived, atom/old_loc, list/atom/old_locs)
 	. = ..()
@@ -245,7 +250,7 @@
 	keys = list(/obj/item/key/church, /obj/item/key/graveyard)
 
 /obj/item/storage/keyring/hand
-	keys = list(/obj/item/key/hand, /obj/item/key/steward, /obj/item/key/tavern, /obj/item/key/church, /obj/item/key/merchant, /obj/item/key/dungeon, /obj/item/key/walls, /obj/item/key/garrison, /obj/item/key/forrestgarrison, /obj/item/key/atarms, /obj/item/key/manor, /obj/item/key/guest)
+	keys = list(/obj/item/key/hand, /obj/item/key/manor, /obj/item/key/steward, /obj/item/key/church, /obj/item/key/merchant, /obj/item/key/dungeon, /obj/item/key/walls, /obj/item/key/garrison, /obj/item/key/forrestgarrison, /obj/item/key/atarms)
 
 /obj/item/storage/keyring/steward
 	keys = list(/obj/item/key/steward, /obj/item/key/vault, /obj/item/key/manor, /obj/item/key/warehouse)
